// src/modules/governance/entities/reg-comp-type.enum.ts
export enum RegCompType {
  AUDIT_SCHEDULER = 'audit_scheduler',
  COMPLIANCE_TRAINING = 'compliance_training',
  DLP_POLICY_ENGINE = 'dlp_policy_engine',
  INSIDER_THREAT_DETECTION = 'insider_threat_detection',
  THIRD_PARTY_RISK = 'third_party_risk',
  CONTINUOUS_ATTESTATION = 'continuous_attestation',
  REGULATORY_CHANGE_MONITOR = 'regulatory_change_monitor',
  AUTOMATED_REPORTING = 'automated_reporting',
  COMPLIANCE_METRICS = 'compliance_metrics',
}